Adds a script to check smokey JSON for a feature
Given a JSON output file from Smokey (`cucumber --format json`) and a
feature name this script will print out a report of that feature's run
and return an exit status; 0 if there were no failures 2 if there were
failures.

This is similar to the script in alphagov/smokey but a bit simpler. It
doesn't log or report on a couple of failures that it should.