Add test that image fallbacks are correctly sized
We had a bug last week that the image fallbacks we had been serving
were unstyled because of PhantomJS's inability to connect to our
assets host using SSLv3.

This bug was present in production for almost a month and nobody
noticed because we have no automated testing of our images.

This commit introduces a very rough check for our image fallbacks.
It asserts that a particular image is approximately a certain size.
I expect this would only catch the most obvious breakages, but it's
better than nothing.