Paths no longer redirect.
After migrating data_sets to resource view in stagecraft, these routes
no longer need to redirect - they are valid and return a 401 when they
are not authorized. There are tests in the app itself for this and so it
seems uncessary to repeat the first feature in this file two more times
to make the point.