The administer dashboards link text has changed
It may change again in future or change back. We should be agnostic
about the text and use a selector. A class or id would be better but
this should work for now and it's debateable whether we should be adding
in classes or ids just for test purposes without agreeing on a
convention.