Fix selector datasets list after perms change
Due to the new feature in the admin app that allows users to select
datasets from a dropdown list, we have given access to all datasets to
users with the dashboard-editor permission, which the smokey user has.
This causes the test to fail because there are 2 datasets with a
datagroup of test, and the selector is ambiguous because it matches 2
elements. By switching to first-child, we will only match one element.

This will still fail if another dataset is added with a datagroup of
test that is listed before test, but I think this fix is good enough for
now.