Fix smokey tests
Update urls:
1. Remove beta from the admin app url.
2. Change spotlight to use the url assigned by gov.uk.
3. Use spotlight subdomain for spotlight assets.

Remove timeout checks:
1. Remove 6 second timeout check for loading spotlight.
2. Remove 4 second timeout check for loading screenshots.

One test scenario is still failing: Image fallbacks look vaguely correct
WebKit no longer trusts SHA1 encrypted certificates.  This is causing
PhantomJS to return a SSL handshake error in Preview.  This is not a
problem in Production as the certificates are SHA2 encrypted.

We're waiting for the certificates to be updated in Preview. When this
happens the test should pass with no futher action required.